Adding Bugs to dominion.c for assignment 2
Smithy | do_smithy():
I changed the for loop condition from “ i < 3” to “i <= 3”. This will make the Smithy card behave incorrectly by making the player draw 4 cards instead of 3.

Adventurer | do_adventurer():
I changed the conditional statement that checks for copper, silver, or gold to only check for copper and silver. This will make the Adventurer card behave incorrectly when the player draws a gold card.

Village | do_village():

I added an extra drawCard(currentPlayer, state) function call to the code for Village. This will give the player an extra card that they aren’t supposed to receive. 

Remodel |do_remodel():

I changed the for loop to begin at “i=1” at the stage where each other player draws a card. This change to the for loops starting position will make it so that every other player minus 1 will draw a card, which is not how Remodel is supposed to work.